D25-20 [Fix] 회원 조회 memberId 추가,sse 도메인 분리
