+++ server/oneyearfourcut/src/main/java/com/codestates/mainproject/oneyearfourcut/domain/alarm/event/AlarmEventListener.java
@@ -1,6 +1,7 @@
 package com.codestates.mainproject.oneyearfourcut.domain.alarm.event;
 
 import com.codestates.mainproject.oneyearfourcut.domain.alarm.service.AlarmService;
+import com.codestates.mainproject.oneyearfourcut.domain.sse.service.SseService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -9,10 +10,11 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class AlarmEventListener {
     private final AlarmService alarmService;
+    private final SseService sseService;
 
     @EventListener
     public void handleAlarmEvent(AlarmEvent event) {
         alarmService.createAlarm(event.getReceiverId(), event.getSenderId(), event.getAlarmType(), event.getGalleryId(), event.getArtworkId());
-        alarmService.send(event.getReceiverId());
+        sseService.send(event.getReceiverId());
     }
 }
