+++ server/oneyearfourcut/src/main/java/com/codestates/mainproject/oneyearfourcut/domain/member/entity/Member.java
@@ -9,7 +9,6 @@ import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.Gallery;
 import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.GalleryStatus;
 import com.codestates.mainproject.oneyearfourcut.domain.member.dto.MemberResponseDto;
 import com.codestates.mainproject.oneyearfourcut.global.auditable.Auditable;
-import com.codestates.mainproject.oneyearfourcut.domain.refreshToken.entity.RefreshToken;
 import lombok.*;
 
 import javax.persistence.*;
@@ -82,6 +81,7 @@ public class Member extends Auditable {
         return MemberResponseDto.builder()
                 .nickname(this.nickname)
                 .profile(this.profile)
+                .memberId(this.memberId)
                 .galleryId(galleryId)
                 .build();
     }
